+++ src/main/java/gift/domain/user/UserService.java
@@ -1,5 +1,6 @@
 package gift.domain.user;
 
+import gift.domain.user.dto.UserDTO;
 import gift.global.jwt.JwtProvider;
 import gift.global.exception.BusinessException;
 import org.springframework.http.HttpStatus;
