+++ src/main/java/gift/domain/user/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package gift.domain.user;
+package gift.domain.user.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
